Changed collection class functionality
Collection class now extends the standard MongoDB Collection class.
This cuts down on unnecessary code to inherit all of the methods of the base class, while still allowing the addition of new methods.

The Database class has been given 2 new methods to enable this and the test class has been updated to reflect the changes